requirements.txt
aiocontextvars==0.2.2
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.5
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.4
boltons==20.2.1
boto3==1.26.149
botocore==1.29.149
Brlapi==0.6.7
cache==1.0.3
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==7.1.2
colorclass==2.2.0
colorlog==5.0.1
command-not-found==0.3
configparser==5.0.2
contextvars==2.4
coverage==5.5
croniter==1.0.12
cryptography==2.3
cupshelpers==1.0
Cython==0.29.23
decorator==5.0.7
defer==1.0.6
distro==1.3.0
distro-info===0.21ubuntu2
dnspython==2.3.0
docopt==0.6.2
entrypoints==0.3
environs==9.3.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fake-useragent==0.1.11
faust==1.10.4
future==0.18.2
fuzzywuzzy==0.17.0
h11==0.14.0
httplib2==0.11.3
idna==2.10
idna-ssl==1.1.0
img2pdf==0.3.1
immutables==0.15
importlib-metadata==4.0.1
iniconfig==1.1.1
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
jmespath==1.0.1
joblib==1.0.1
kafka-python==1.4.7
keyring==17.1.1
keyrings.alt==3.1.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
louis==3.8.0
lxml==4.9.2
macaroonbakery==1.2.1
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==3.11.1
mccabe==0.6.1
mode==4.3.2
msgpack==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.33
netifaces==0.10.4
nltk==3.6.2
numpy==1.19.5
oauthlib==2.1.0
olefile==0.46
openpyxl==3.1.2
opentracing==1.3.0
outcome==1.2.0
packaging==20.9
pandas==1.3.5
parso==0.8.2
pathlib==1.0.1
pbr==5.11.1
pdfkit==0.6.1
pdfrw==0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipreqs==0.4.13
pluggy==0.13.1
prompt-toolkit==3.0.18
protobuf==3.20.3
psycopg2-binary==2.9.6
ptyprocess==0.7.0
py==1.10.0
PyAutoIt==0.6.5
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.8.1
PyGObject==3.32.0
pyhtml2pdf==0.0.6
PyJWT==1.7.0
pylint==2.6.0
pymacaroons==0.13.0
pymongo==4.4.1
PyNaCl==1.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyRFC3339==1.1
PySocks==1.7.1
pytest==6.2.3
pytest-cov==2.11.1
python-apt===1.8.5-ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.34
python-dotenv==0.17.0
python-slugify==4.0.1
pytz==2018.9
pyxattr==0.6.1
pyxdg==0.25
PyYAML==3.13
redis==3.5.3
regex==2021.4.4
reportlab==3.5.67
requests==2.25.1
requests-mock==1.9.1
requests-unixsocket==0.1.5
robinhood-aiokafka==1.1.6
s3transfer==0.6.1
SecretStorage==2.3.1
selenium==4.9.0
simplejson==3.16.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
system-service==0.3
systemd-python==234
terminaltables==3.1.0
testresources==2.0.1
text-unidecode==1.3
toml==0.10.1
tqdm==4.60.0
traitlets==4.3.3
trio==0.22.0
trio-websocket==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
ubuntu-advantage-tools==19.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.2.0
urllib3==1.26.4
usb-creator==0.3.3
uuid==1.30
venusian==1.2.0
wadllib==1.3.3
wcwidth==0.2.5
webdriver-manager==3.8.6
wrapt==1.12.1
wsproto==1.2.0
xkit==0.0.0
yarg==0.1.9
yarl==1.6.3
youtube-dl==2019.1.17
zipp==3.4.1
